+++ swe.py
@@ -1,107 +1,64 @@
+import countries_codes
 import requests
+import pprint
+import urllib
 import os
-import re
-import time
-from bs4 import BeautifulSoup
-import warnings
 
-def get_token():
-    return os.environ.get('FINDWORK_TOKEN')
+class Backend:        
+    #Getting the Adzuna id, key
+    def get_token(self):
+        return os.environ.get('ADZUNA_ID'), os.environ.get('ADZUNA_KEY')
 
-def prepare_request(token):
-    url = 'https://findwork.dev/api/jobs/'
-    headers = {'Authorization': 'Token ' + token}
-    params = {}
-    location =  input('Enter location ["Remote" for remote, or leave blank if none]: ')
+    #Returning the ISO3 code of a given country
+    def get_country_key(self, country_num):
+        return countries_codes.countries_dict[country_num]
 
-    if "remote" in location.lower():
-        params['remote'] = True
-    else:
-        params['location'] = location
+    #Prepering the url and parameters
+    def prepare_request(self, id, key, country_num, location):
+        params = {}
+        country_code = self.get_country_key(country_num)
+        page_num = 1
+        params['app_id'] = id
+        params['app_key'] = key
+        params['results_per_page'] = 100
+        params['what_or'] = "software developer programmer designer ML security data product"
+        params['title_only'] = "intern"
+        params['where'] =  location
+        params['sort_by'] = "relevance"
 
-    params['search'] = input('Enter search term [Leave blank if none]: ')
-    return url, headers, params
+        url = f'https://api.adzuna.com/v1/api/jobs/{country_code}/search/{page_num}?'
+        params = urllib.parse.urlencode(params, quote_via=urllib.parse.quote)
+        return url, params 
 
-def get_results(url, headers, params):
-    data = requests.get(url,
-                        headers = headers,
-                        params = params).json()
-    return data.get('results'), data.get('next')
-
-def role_is_intern(role):
-    match = re.compile('intern(.*)', re.IGNORECASE)
-    exclude = re.compile('^(ation)?al')
-    search = match.search(role)
-    if not search:
-        return False
-    if exclude.search(search.group(1)):
-        return False
-    return True
-
-def is_tech_role(role):
-    match = re.compile('^.*(software|developer|programmer|designer|ML|security|data|product|engineer|engineering).*$', re.IGNORECASE)
-    search = match.search(role)
-    if not search:
-        return False
-    return True
+    #Get a request
+    def get_results(self, url, params):
+        data = requests.get(url, params = params).json()
+        return data['results']
 
 if __name__ == '__main__':
-    warnings.filterwarnings("ignore", category=UserWarning, module='bs4')
-    TOKEN = get_token()
-    url, headers, params = prepare_request(TOKEN)
-    count = 0
 
-    #MOVE TO FRONT
+    swe_api = Backend()
 
-    while url is not None:
+   # warnings.filterwarnings("ignore", category=UserWarning, module='bs4')
+    ADZUNA_ID, ADZUNA_KEY = swe_api.get_token()
+    url, params = swe_api.prepare_request(ADZUNA_ID, ADZUNA_KEY, '15', "san francisco")
+    results = swe_api.get_results(url, params)
 
-        results, next = get_results(url, headers, params)
-        
-        for result in results:
-            role = result.get('role')
-            if role_is_intern(role) and is_tech_role(role):
-                print(role)
-                count += 1
-                soup = BeautifulSoup(result.get('text'), features="html.parser")
+    for job in results:
+        title = job.get('title')
+        print(title)
+
+#               soup = BeautifulSoup(result.get('text'), features="html.parser")
                 
-                if result.get('remote') is True:
-                    print(result.get('company_name') +  " (Remote)")
-                elif result.get('location') is not None:
-                    if result.get('location')[-1] is " ":
-                        result['location'] = result.get('location')[:-2]
-                    print(result.get('company_name') + " (" + result.get('location') + ")")
-                else:
-                    print(result.get('company_name'))
+               # if result.get('remote') is True:
+                #    print(result.get('company_name') +  " (Remote)")
+                #elif result.get('location') is not None:
+                #    if result.get('location')[-1] is " ":
+                #        result['location'] = result.get('location')[:-2]
+                  #  print(result.get('company_name') + " (" + result.get('location') + ")")
+                #else:
+                 #   print(result.get('company_name'))
                 
                 #Prints the description 
-                print(soup.get_text()[:260] + "...")
+                #print(soup.get_text()[:260] + "...")
                 ##Ask "yes, next etc" and add to database
-
-        url = next
-        time.sleep(1) 
-
-    print(count)
-
-
-
-
-
-
-
-    countries_dict = {
-    "1" : "Austria",
-    "2" : "Brazil",
-    "3":  "Canada",
-    "4" : "France",
-    "5" : "Germany",
-    "6" : "India",
-    "7" : "Italy",
-    "8" : "Netherlands",
-    "9" : "New Zealand",
-    "10" : "Poland",
-    "11" : "Russia",
-    "12" : "Singapore",
-    "13" : "South Africa",
-    "14" : "United Kingdom",
-    "15" : "us",
-}
